add(test/verify): moved from pkg/verify
