+++ test/verify/imagestreams.go
